# perf: pre-create `UnionType` objects