--- a/frappe/utils/response.py
+++ b/frappe/utils/response.py
@@ -8,7 +8,9 @@ import mimetypes
 import os
 import sys
 import uuid
+from collections.abc import Iterable
 from pathlib import Path
+from re import Match
 from typing import TYPE_CHECKING
 from urllib.parse import quote
 
@@ -29,7 +31,8 @@ from frappe.utils.local import LocalProxy, WerkzeugLocalProxy
 if TYPE_CHECKING:
 	from frappe.core.doctype.file.file import File
 
-LocalProxyType = LocalProxy | WerkzeugLocalProxy
+LocalProxyTypes = LocalProxy | WerkzeugLocalProxy
+DateOrTimeTypes = datetime.date | datetime.datetime | datetime.time
 
 
 def report_error(status_code):
@@ -212,19 +215,14 @@ def _make_logs_v2():
 
 def json_handler(obj):
 	"""serialize non-serializable data for json"""
-	from collections.abc import Iterable
-	from re import Match
 
-	if isinstance(obj, datetime.date | datetime.datetime | datetime.time):
+	if isinstance(obj, DateOrTimeTypes):
 		return str(obj)
 
 	elif isinstance(obj, datetime.timedelta):
 		return format_timedelta(obj)
 
-	elif isinstance(obj, decimal.Decimal):
-		return float(obj)
-
-	elif isinstance(obj, LocalProxyType):
+	elif isinstance(obj, LocalProxyTypes):
 		return str(obj)
 
 	elif hasattr(obj, "__json__"):
@@ -232,6 +230,9 @@ def json_handler(obj):
 
 	elif isinstance(obj, Iterable):
 		return list(obj)
+
+	elif isinstance(obj, decimal.Decimal):
+		return float(obj)
 
 	elif isinstance(obj, Match):
 		return obj.string