This is apparently expensive in functions called repeatedly.

### Before

```py
In [8]: %timeit dic = d.as_dict(convert_dates_to_str=True)
1.18 ms ± 13 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [11]: %timeit json.dumps(dic, default=json_handler, separators=(",", ":"))
378 µs ± 924 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

### After

```py
In [7]: %timeit dic = d.as_dict(convert_dates_to_str=True)
740 µs ± 14.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [12]: %timeit json.dumps(dic, default=json_handler, separators=(",", ":"))
316 µs ± 8 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```